+++ CMakeLists.txt
@@ -3,7 +3,7 @@ cmake_minimum_required (VERSION 2.8)
 include(ExternalProject)
 
 project(pHash)
-
+set(CMAKE_MACOSX_RPATH 1)
 set(pHash_VERSION_MAJOR 1)
 set(pHash_VERSION_MINOR 0)
 
@@ -46,18 +46,31 @@ add_subdirectory(bindings)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BUILD_TYPE}")
 
 link_directories(/usr/local/lib)
-
-add_library(pHash SHARED ${MMAN_FILE} ${DIRENT_FILE} src/pHash.cpp )
+include_directories(/usr/local/include)
 
 
 set(LIBS_DEPS 			png jpeg tiff pthread
 				)
 
+find_library(LIBMPG123 mpg123)
+if(LIBMPG123)
+	message("libmpg123 found at ${LIBMPG123}")
+	set(HAVE_LIBMPG123 1)
+	list(APPEND LIBS_DEPS ${LIBMPG123})
+endif()
 
 if(HAVE_AUDIO_HASH)
-	add_library(pHash src/audiophash.cpp)
-	set(LIBS_DEPS "${LIBS_DEPS} sndfile samplerate vorbis vorbisenc ogg")
+	set(EXTRA_SRC src/audiophash.cpp src/ph_fft.cpp)
+	set(LIBS_DEPS ${LIBS_DEPS} sndfile samplerate vorbis vorbisenc ogg)
 endif()
+
+if(HAVE_VIDEO_HASH)
+	list(APPEND EXTRA_SRC src/cimgffmpeg.cpp)
+	set(LIBS_DEPS ${LIBS_DEPS} avcodec avformat swscale avutil)
+endif()
+
+add_library(pHash SHARED ${MMAN_FILE} ${DIRENT_FILE} src/pHash.cpp ${EXTRA_SRC})
+
 target_link_libraries(pHash	${LIBS_DEPS})
 
 install(TARGETS pHash DESTINATION lib)
@@ -65,6 +78,7 @@ install(FILES pHash.h DESTINATION include)
 
 set(EXAMPLEDIR examples)
 include_directories(${CMAKE_CURRENT_SOURCE_DIR})
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/src)
 
 if(HAVE_IMAGE_HASH)
 add_executable(TestDCT "${EXAMPLEDIR}/test_imagephash.cpp")
@@ -74,10 +88,6 @@ install(TARGETS TestDCT DESTINATION bin)
 add_executable(TestRadish "${EXAMPLEDIR}/test_radish.cpp")
 target_link_libraries (TestRadish pHash)
 install(TARGETS TestRadish DESTINATION bin)
-
-#add_executable(TestBMB "${EXAMPLEDIR}/test_bmbimagehash.cpp")
-#target_link_libraries (TestBMB pHash)
-#install(TARGETS TestBMB DESTINATION bin)
 
 add_executable(TestMH "${EXAMPLEDIR}/test_mhimagehash.cpp")
 target_link_libraries (TestMH pHash)
