+++ src/cimgffmpeg.h
@@ -35,9 +35,9 @@
 #define __STDC_CONSTANT_MACROS
 
 extern "C" {
-	#include "./libavformat/avformat.h"
-	#include "./libavcodec/avcodec.h"
-	#include "./libswscale/swscale.h"
+	#include <libavformat/avformat.h>
+	#include <libavcodec/avcodec.h>
+	#include <libswscale/swscale.h>
 }
 
 using namespace cimg_library;
