+++ src/pHash.h.cmake
@@ -46,6 +46,7 @@
 #cmakedefine HAVE_IMAGE_HASH
 #cmakedefine HAVE_AUDIO_HASH
 #cmakedefine HAVE_VIDEO_HASH
+#cmakedefine HAVE_LIBMPG123
 
 #define PACKAGE_STRING "${CMAKE_PROJECT_NAME}"
 
@@ -86,36 +87,6 @@ typedef  int64_t  long64;
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-struct BinHash 
-{
-	uint8_t *hash;
-	uint32_t bytelength;
-	uint32_t byteidx; // used by addbit()
-	uint8_t bitmask;  // used by addbit()
-
-	/*
-	 * add a single bit to hash. the bits are 
-	 * written from left to right.
-	 */
-	int addbit(uint8_t bit)
-	{
-		if (bitmask == 0) 
-		{
-			bitmask = 128; // reset bitmask to "10000000"
-			byteidx++;     // jump to next byte in array
-		}
-
-		if (byteidx >= bytelength) return -1;
-		
-		if (bit == 1) *(hash + byteidx) |= bitmask;
-		bitmask >>=1;
-		return 0;
-	}	
-};
-
-BinHash* _ph_bmb_new(uint32_t bytelength);
-void ph_bmb_free(BinHash *binHash);
 
 /*! /brief Radon Projection info
  */
@@ -261,7 +232,6 @@ static CImg<float>* ph_dct_matrix(const int N);
  */
 int ph_dct_imagehash(const char* file,ulong64 &hash);
 
-int ph_bmb_imagehash(const char *file, uint8_t method, BinHash **ret_hash);
 
 #ifdef HAVE_VIDEO_HASH
 static CImgList<uint8_t>* ph_getKeyFramesFromVideo(const char *filename);
