+++ src/ph_fft.cpp
@@ -26,17 +26,10 @@
 
 #include "ph_fft.h"
 
-complex double polar_to_complex(const double r, const double theta)
-{
-    complex double result;
-    result = r*cos(theta) + r*sin(theta)*I;
 
-    return result;
-}
-
-void fft_calc(const int N,const double *x,complex double *X,complex double *P,const int step,const complex double *twids)
+void fft_calc(const int N,const double *x,complex<double> *X,complex<double> *P,const int step,const complex<double> *twids)
 {
-    complex double *S = P + N/2;
+    complex<double> *S = P + N/2;
     if (N == 1){
 	X[0] = x[0];
 	return;
@@ -55,20 +48,20 @@ void fft_calc(const int N,const double *x,complex double *X,complex double *P,co
 }
 
 
-int fft(double *x, int N, complex double *X)
+int fft(double *x, int N, complex<double> *X)
 {
 
-    complex double *twiddle_factors = (complex double*)malloc(sizeof(complex double)*(N/2));
-    complex double *Xt = (complex double*)malloc(sizeof(complex double)*N);
+    complex<double> *twiddle_factors = new complex<double>[N/2];
+    complex<double> *Xt = new complex<double>[N];
 
     int k;
     for (k=0;k<N/2;k++){
-	twiddle_factors[k] = polar_to_complex(1.0, 2.0*PI*k/N);
+	twiddle_factors[k] = polar(1.0, 2.0*M_PI*k/N);
     }
     fft_calc(N, x, X, Xt, 1, twiddle_factors);
 
-    free(twiddle_factors);
-    free(Xt);
+    delete [] (twiddle_factors);
+    delete [] (Xt);
 
     return 0;
